refactor: Extracted PageFactory from ScraperBaseHandler
