+++ JobScraper.Web/Features/JobOffers/Scrape/Logic/Common/DetailsScraperBaseHandler.cs
@@ -68,7 +68,7 @@ public abstract partial class DetailsScraperBaseHandler<TScrapeCommand>(
             }
             finally
             {
-                Dispose();
+                await DisposeAsync();
             }
         }
